Treat all JIT buffer parameters as non-linear parameters
This commit addresses very large kernels that were generated by the
JIT for smaller buffers. This was occuring because of how we were
estimating the JIT parameter sizes.

When targeting linear kernels we only pass the pointer to the kernel
instead of the entire dims and stride arrays. This generates smaller
parameter sets so we calculated kernels based on the pointer size
instead of the Param<T> size.

The issue occured when this kernel came in below the threshold for
evaluation but when it was used as a node to another operation, the
generated kernel became a non-linear kernel requiring us to pass
the Param<T> object for each buffer. This causes the size to go past
the limit of the parameters supported by CUDA and compilation fails.

This commit treats all kernels as non-linear kernel when calculating
the parameter sizes.

Fixes #2436 #2389

(cherry picked from commit 9007742719f6decfb705b49781d6d2aaf23402de)

+++ test/sparse_common.hpp
@@ -142,11 +142,10 @@ void sparseTransposeTester(const int m, const int n, const int k, int factor, do
     ASSERT_NEAR(0, calc_norm(imag(dRes3), imag(sRes3)), eps);
 }
 
-template<typename T> static
-void convertCSR(const int M, const int N, const float ratio, int targetDevice=-1)
-{
-    if (targetDevice>=0)
-        af::setDevice(targetDevice);
+template<typename T>
+static void convertCSR(const int M, const int N, const double ratio,
+                       int targetDevice = -1) {
+    if (targetDevice >= 0) af::setDevice(targetDevice);
 
     if (noDoubleTests<T>()) return;
 #if 1
